Disable plugin for images bigger than 512x512
The Gaussian Blur filter produces artifacts on the border of the tiles.
Disabling tiling for a given filter is not possible with current Quill, so
the only possible option at the moment is using pictures so small that they
don't use tiling.

This patch disables the widget for images larger than 512x512.